Refactor map loading
Constructing gems should not be happening in the map parser, separated from the board class where Gems get created. Rework the map loading so that now only indexes are stored. These get converted to types of Gems in the Board class. Also define a default starting map. This gets around the problem of randomly generating a grid without matches.